Move CreateTeam module
